api-bancaria/pyproject.toml
[project]
name = "api-bancaria"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitor",email = "vitordeazambuja@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.127.1,<0.128.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
